requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
Pillow==9.5.0
psycopg-binary==3.1.9
pytz==2023.3
sqlparse==0.4.4
